+++ src/PendingObjects/TestCall.php
@@ -148,6 +148,9 @@ final class TestCall
             ? $conditionOrMessage
             : $message;
 
+        /** @var callable(): bool $condition */
+        $condition = $condition->bindTo(null);
+
         $this->testCaseFactory
             ->chains
             ->addWhen($condition, Backtrace::file(), Backtrace::line(), 'markTestSkipped', [$message]);